refactor all into stages
