+++ docling/experimental/pipeline/threaded_layout_vlm_pipeline.py
@@ -33,12 +33,12 @@ from docling.experimental.datamodel.threaded_layout_vlm_pipeline_options import 
     ThreadedLayoutVlmPipelineOptions,
 )
 from docling.models.base_model import BaseVlmPageModel
-from docling.models.layout_model import LayoutModel
-from docling.models.vlm_models.api_vlm_model import ApiVlmModel
-from docling.models.vlm_models.hf_transformers_model import (
+from docling.models.stages.layout.layout_model import LayoutModel
+from docling.models.vlm_pipeline_models.api_vlm_model import ApiVlmModel
+from docling.models.vlm_pipeline_models.hf_transformers_model import (
     HuggingFaceTransformersVlmModel,
 )
-from docling.models.vlm_models.mlx_model import HuggingFaceMlxModel
+from docling.models.vlm_pipeline_models.mlx_model import HuggingFaceMlxModel
 from docling.pipeline.base_pipeline import BasePipeline
 from docling.pipeline.standard_pdf_pipeline import (
     ProcessingResult,
@@ -162,7 +162,7 @@ class ThreadedLayoutVlmPipeline(BasePipeline):
                     vlm_options=vlm_options,
                 )
             elif vlm_options.inference_framework == InferenceFramework.VLLM:
-                from docling.models.vlm_models.vllm_model import VllmVlmModel
+                from docling.models.vlm_pipeline_models.vllm_model import VllmVlmModel
 
                 self.vlm_model = VllmVlmModel(
                     enabled=True,
