+++ docling/models/extraction/nuextract_transformers_model.py
