+++ docling/models/stages/ocr/easyocr_model.py
