+++ docling/models/stages/ocr/ocr_mac_model.py
