+++ docling/models/stages/ocr/rapid_ocr_model.py
