+++ docling/models/stages/ocr/tesseract_ocr_cli_model.py
