+++ docling/models/stages/ocr/tesseract_ocr_model.py
