+++ docling/models/stages/picture_classifier/document_picture_classifier.py
