+++ docling/models/stages/picture_description/picture_description_api_model.py
