+++ docling/models/stages/picture_description/picture_description_vlm_model.py
