+++ docling/models/stages/table_structure/table_structure_model.py
