+++ docling/models/vlm_pipeline_models/hf_transformers_model.py
