+++ docling/models/vlm_pipeline_models/vllm_model.py
