+++ docling/utils/model_downloader.py
@@ -15,15 +15,19 @@ from docling.datamodel.vlm_model_specs import (
     SMOLDOCLING_MLX,
     SMOLDOCLING_TRANSFORMERS,
 )
-from docling.models.code_formula_model import CodeFormulaModel
-from docling.models.document_picture_classifier import DocumentPictureClassifier
-from docling.models.layout_model import LayoutModel
-from docling.models.ocr.easyocr_model import EasyOcrModel
-from docling.models.ocr.rapid_ocr_model import RapidOcrModel
-from docling.models.picture_description.picture_description_vlm_model import (
+from docling.models.stages.code_formula.code_formula_model import CodeFormulaModel
+from docling.models.stages.layout.layout_model import LayoutModel
+from docling.models.stages.ocr.easyocr_model import EasyOcrModel
+from docling.models.stages.ocr.rapid_ocr_model import RapidOcrModel
+from docling.models.stages.picture_classifier.document_picture_classifier import (
+    DocumentPictureClassifier,
+)
+from docling.models.stages.picture_description.picture_description_vlm_model import (
     PictureDescriptionVlmModel,
 )
-from docling.models.table_structure.table_structure_model import TableStructureModel
+from docling.models.stages.table_structure.table_structure_model import (
+    TableStructureModel,
+)
 from docling.models.utils.hf_model_download import download_hf_model
 
 _log = logging.getLogger(__name__)
